gen1/skills_mcp/pyproject.toml
[project]
name = "skills-mcp"
version = "1.0.0"
description = "Standalone production-grade Skills MCP Server"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "crewai>=0.100.0",
    "crewai-tools[mcp]>=0.100.0",
]

[project.scripts]
skills-mcp       = "mcp_server.__main__:main"   # stdio
skills-api       = "app:main"                    # HTTP

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
